# Conflicts:
#	admin/configuration.php
#	include/class/classProduct.php
#	include/func/func.order.php
#	include/order_edit.php
#	modules/External_Product_Verification/include/classExternalVerificationBatches.php
#	modules/XPDF/func.php
#	testDF78KL54AS01.php

+++ include/order_status_notifications.php
@@ -1,40 +1,43 @@
 <?php
 /*****************************************************************************\
-+-----------------------------------------------------------------------------+
-| X-Cart                                                                      |
-| Copyright (c) 2001-2011 Ruslan R. Fazliev <rrf@rrf.ru>                      |
-| All rights reserved.                                                        |
-+-----------------------------------------------------------------------------+
-| PLEASE READ  THE FULL TEXT OF SOFTWARE LICENSE AGREEMENT IN THE "COPYRIGHT" |
-| FILE PROVIDED WITH THIS DISTRIBUTION. THE AGREEMENT TEXT IS ALSO AVAILABLE  |
-| AT THE FOLLOWING URL: http://www.x-cart.com/license.php                     |
-|                                                                             |
-| THIS  AGREEMENT  EXPRESSES  THE  TERMS  AND CONDITIONS ON WHICH YOU MAY USE |
-| THIS SOFTWARE   PROGRAM   AND  ASSOCIATED  DOCUMENTATION   THAT  RUSLAN  R. |
-| FAZLIEV (hereinafter  referred to as "THE AUTHOR") IS FURNISHING  OR MAKING |
-| AVAILABLE TO YOU WITH  THIS  AGREEMENT  (COLLECTIVELY,  THE  "SOFTWARE").   |
-| PLEASE   REVIEW   THE  TERMS  AND   CONDITIONS  OF  THIS  LICENSE AGREEMENT |
-| CAREFULLY   BEFORE   INSTALLING   OR  USING  THE  SOFTWARE.  BY INSTALLING, |
-| COPYING   OR   OTHERWISE   USING   THE   SOFTWARE,  YOU  AND  YOUR  COMPANY |
-| (COLLECTIVELY,  "YOU")  ARE  ACCEPTING  AND AGREEING  TO  THE TERMS OF THIS |
-| LICENSE   AGREEMENT.   IF  YOU    ARE  NOT  WILLING   TO  BE  BOUND BY THIS |
-| AGREEMENT, DO  NOT INSTALL OR USE THE SOFTWARE.  VARIOUS   COPYRIGHTS   AND |
-| OTHER   INTELLECTUAL   PROPERTY   RIGHTS    PROTECT   THE   SOFTWARE.  THIS |
-| AGREEMENT IS A LICENSE AGREEMENT THAT GIVES  YOU  LIMITED  RIGHTS   TO  USE |
-| THE  SOFTWARE   AND  NOT  AN  AGREEMENT  FOR SALE OR FOR  TRANSFER OF TITLE.|
-| THE AUTHOR RETAINS ALL RIGHTS NOT EXPRESSLY GRANTED BY THIS AGREEMENT.      |
-|                                                                             |
-| The Initial Developer of the Original Code is Ruslan R. Fazliev             |
-| Portions created by Ruslan R. Fazliev are Copyright (C) 2001-2011           |
-| Ruslan R. Fazliev. All Rights Reserved.                                     |
-+-----------------------------------------------------------------------------+
-\*****************************************************************************/
+ * +-----------------------------------------------------------------------------+
+ * | X-Cart                                                                      |
+ * | Copyright (c) 2001-2011 Ruslan R. Fazliev <rrf@rrf.ru>                      |
+ * | All rights reserved.                                                        |
+ * +-----------------------------------------------------------------------------+
+ * | PLEASE READ  THE FULL TEXT OF SOFTWARE LICENSE AGREEMENT IN THE "COPYRIGHT" |
+ * | FILE PROVIDED WITH THIS DISTRIBUTION. THE AGREEMENT TEXT IS ALSO AVAILABLE  |
+ * | AT THE FOLLOWING URL: http://www.x-cart.com/license.php                     |
+ * |                                                                             |
+ * | THIS  AGREEMENT  EXPRESSES  THE  TERMS  AND CONDITIONS ON WHICH YOU MAY USE |
+ * | THIS SOFTWARE   PROGRAM   AND  ASSOCIATED  DOCUMENTATION   THAT  RUSLAN  R. |
+ * | FAZLIEV (hereinafter  referred to as "THE AUTHOR") IS FURNISHING  OR MAKING |
+ * | AVAILABLE TO YOU WITH  THIS  AGREEMENT  (COLLECTIVELY,  THE  "SOFTWARE").   |
+ * | PLEASE   REVIEW   THE  TERMS  AND   CONDITIONS  OF  THIS  LICENSE AGREEMENT |
+ * | CAREFULLY   BEFORE   INSTALLING   OR  USING  THE  SOFTWARE.  BY INSTALLING, |
+ * | COPYING   OR   OTHERWISE   USING   THE   SOFTWARE,  YOU  AND  YOUR  COMPANY |
+ * | (COLLECTIVELY,  "YOU")  ARE  ACCEPTING  AND AGREEING  TO  THE TERMS OF THIS |
+ * | LICENSE   AGREEMENT.   IF  YOU    ARE  NOT  WILLING   TO  BE  BOUND BY THIS |
+ * | AGREEMENT, DO  NOT INSTALL OR USE THE SOFTWARE.  VARIOUS   COPYRIGHTS   AND |
+ * | OTHER   INTELLECTUAL   PROPERTY   RIGHTS    PROTECT   THE   SOFTWARE.  THIS |
+ * | AGREEMENT IS A LICENSE AGREEMENT THAT GIVES  YOU  LIMITED  RIGHTS   TO  USE |
+ * | THE  SOFTWARE   AND  NOT  AN  AGREEMENT  FOR SALE OR FOR  TRANSFER OF TITLE.|
+ * | THE AUTHOR RETAINS ALL RIGHTS NOT EXPRESSLY GRANTED BY THIS AGREEMENT.      |
+ * |                                                                             |
+ * | The Initial Developer of the Original Code is Ruslan R. Fazliev             |
+ * | Portions created by Ruslan R. Fazliev are Copyright (C) 2001-2011           |
+ * | Ruslan R. Fazliev. All Rights Reserved.                                     |
+ * +-----------------------------------------------------------------------------+
+ * \*****************************************************************************/
 
 #
 # $Id: order_status_notifications.php, v 1.0.0 2011/10/18 12:50:09 kate Exp $
 #
 
-if ( !defined('XCART_SESSION_START') ) { header('Location: ../'); die('Access denied'); }
+if (!defined('XCART_SESSION_START')) {
+    header('Location: ../');
+    die('Access denied');
+}
 
 x_session_register('selected_status');
 
@@ -53,53 +56,58 @@ if (isset($status)) {
 if ($REQUEST_METHOD == 'POST') {
     if ($mode == 'update') {
         if (is_array($update)) {
-            
-            $top_message['content'] = '';
-            
-            if (empty($update['customer_subject'])) {
-                $top_message['content'] .= func_get_langvar_by_name('err_empty_customer_mail_subject');
+            if (!empty($update['customer_subject'])) {
+                db_query('DELETE FROM ' . $sql_tbl['order_status_notifications'] . ' WHERE code = "' . $selected_status . '"');
+                foreach ($update['customer_subject'] as $idx => $value) {
+
+                    $top_message['content'] = '';
+
+                    if (empty($update['customer_subject'])) {
+                        $top_message['content'] .= func_get_langvar_by_name('err_empty_customer_mail_subject');
+                    }
+
+                    if (empty($update['copy_subject'])) {
+                        $top_message['content'] .= func_get_langvar_by_name('err_empty_copy_mail_subject');
+                    }
+
+                    if (empty($update['customer_subject']) || empty($update['copy_subject'])) {
+                        $top_message['type'] == 'E';
+                        func_header_location('order_status_notifications.php');
+                    }
+
+
+
+                    $query = array(
+                        'customer_subject' => mysql_real_escape_string($update['customer_subject'][$idx]),
+                        'copy_subject' => mysql_real_escape_string($update['copy_subject'][$idx]),
+                        'email_body' => $update['email_body'][$idx],
+                        'enabled' => (isset($update['enabled'][$idx]) && $update['enabled'][$idx] == 'Y') ? 'Y' : 'N',
+                        'customer_attach_pdf_invoice' => (isset($update['customer_attach_pdf_invoice'][$idx]) && $update['customer_attach_pdf_invoice'][$idx] == 'Y') ? 'Y' : 'N',
+                        'admin_attach_pdf_invoice' => (isset($update['admin_attach_pdf_invoice'][$idx]) && $update['admin_attach_pdf_invoice'][$idx] == 'Y') ? 'Y' : 'N',
+                        'code' => $selected_status,
+                        'number' => $idx,
+                    );
+
+                    func_array2insert('order_status_notifications', $query);
+                }
             }
-            
-            if (empty($update['copy_subject'])) {
-                $top_message['content'] .= func_get_langvar_by_name('err_empty_copy_mail_subject');
-            }
-
-            if (empty($update['customer_subject']) || empty($update['copy_subject'])) {
-                $top_message['type'] == 'E';
-                func_header_location('order_status_notifications.php');
-            }
-
-            db_query('DELETE FROM ' . $sql_tbl['order_status_notifications'] . ' WHERE code = "' . $selected_status . '"');
-
-            $query = array(
-                'customer_subject'  => mysql_real_escape_string($update['customer_subject']),
-                'copy_subject'      => mysql_real_escape_string($update['copy_subject']),
-                'email_body'        => $update['email_body'],
-                'enabled'           => (isset($update['enabled']) && $update['enabled'] == 'Y') ? 'Y' : 'N',
-                'customer_attach_pdf_invoice'           => (isset($update['customer_attach_pdf_invoice']) && $update['customer_attach_pdf_invoice'] == 'Y') ? 'Y' : 'N',
-                'admin_attach_pdf_invoice'           => (isset($update['admin_attach_pdf_invoice']) && $update['admin_attach_pdf_invoice'] == 'Y') ? 'Y' : 'N',
-                'code'              => $selected_status,
-            );
-
-            func_array2insert('order_status_notifications', $query);
         }
     }
     func_header_location('order_status_notifications.php');
 }
+global $xcart_dir;
+require_once $xcart_dir . "/include/class/classOrderStatusNotification.php";
+$aOrderNotifications = classOrderStatusNotification::getOrderStatusNotificationsByCode($selected_status);
 
-$osn_settings = func_query_first('SELECT code, customer_subject, copy_subject, email_body, enabled, customer_attach_pdf_invoice, admin_attach_pdf_invoice'
-    . ' FROM ' . $sql_tbl['order_status_notifications']
-    . ' WHERE code = "' . $selected_status . '"');
-
-$smarty->assign('osn_settings', $osn_settings);
+$smarty->assign('aOrderNotifications', $aOrderNotifications);
 
 $status_types = array(
-    'CB'    => func_get_langvar_by_name('lbl_cust_bus_payment_status'),
-    'DC'    => func_get_langvar_by_name('lbl_distr_cust_shipping_status'),
-    'BD'    => func_get_langvar_by_name('lbl_bus_distr_payment_status'),
+    'CB' => func_get_langvar_by_name('lbl_cust_bus_payment_status'),
+    'DC' => func_get_langvar_by_name('lbl_distr_cust_shipping_status'),
+    'BD' => func_get_langvar_by_name('lbl_bus_distr_payment_status'),
 //    'CA'    => 'Currently assigned to',
-    'RU'    => 'REF TO US status',
-    'PO'    => 'Check transit status',
+    'RU' => 'REF TO US status',
+    'PO' => 'Check transit status',
 );
 
 $smarty->assign('status_types', $status_types);
